Magnetic Mouse - trackpad+magnetic forces
This is a rudimentary implementation of the magnetic mouse that gets pulled inside the target box whenever it is near one. The ad hoc mouse movements are controlled via the laptop trackpad. But we the mouse pointer reaches near any target, it gets pulled inside the target box.

The code is simplistic and needs more changes to get better results.